+++ playground/Playground/RealisticUnitOfWork_GeneratedResolutionGraph.cs
@@ -1,0 +1,354 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using DryIoc;
+
+namespace RealisticUnitOfWork
+{
+    internal static class GeneratedResolutionGraph
+    {
+        static FactoryDelegate GetFactory()
+        {
+            return (FactoryDelegate)((IResolverContext r) => //$
+                ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                    164,
+                    (FactoryDelegate)((IResolverContext r) => //$
+                        new R(
+                            default(Single1)/* (!) Please provide the non-default value for the constant */,
+                            default(Single2)/* (!) Please provide the non-default value for the constant */,
+                            ((Scoped1)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                165,
+                                (FactoryDelegate)((IResolverContext r) => //$
+                                    new Scoped1(
+                                        default(Single12)/* (!) Please provide the non-default value for the constant */,
+                                        default(SingleObj12)/* (!) Please provide the non-default value for the constant */,
+                                        ((ScopedFac12)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                            183,
+                                            (FactoryDelegate)((IResolverContext r) => //$
+                                                default(Func<Scoped13, Single1, SingleObj13, ScopedFac12>)/* (!) Please provide the non-default value for the constant */.Invoke(
+                                                    ((Scoped13)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                        187,
+                                                        (FactoryDelegate)((IResolverContext r) => //$
+                                                            new Scoped13(
+                                                                default(Single1)/* (!) Please provide the non-default value for the constant */,
+                                                                ((Scoped14)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                    197,
+                                                                    (FactoryDelegate)((IResolverContext r) => //$
+                                                                        new Scoped14()),
+                                                                    r)))),
+                                                        r)),
+                                                    default(Single1)/* (!) Please provide the non-default value for the constant */,
+                                                    default(SingleObj13)/* (!) Please provide the non-default value for the constant */)),
+                                            r)),
+                                        new Trans12(
+                                            new Trans13(
+                                                default(Single14)/* (!) Please provide the non-default value for the constant */,
+                                                new Trans14()),
+                                            default(Single13)/* (!) Please provide the non-default value for the constant */,
+                                            default(SingleObj13)/* (!) Please provide the non-default value for the constant */),
+                                        default(Single1)/* (!) Please provide the non-default value for the constant */,
+                                        default(SingleObj1)/* (!) Please provide the non-default value for the constant */,
+                                        ((Scoped12)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                            177,
+                                            (FactoryDelegate)((IResolverContext r) => //$
+                                                new Scoped12(
+                                                    default(Single13)/* (!) Please provide the non-default value for the constant */,
+                                                    default(SingleObj13)/* (!) Please provide the non-default value for the constant */,
+                                                    ((Scoped13)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                        187,
+                                                        (FactoryDelegate)((IResolverContext r) => //$
+                                                            new Scoped13(
+                                                                default(Single1)/* (!) Please provide the non-default value for the constant */,
+                                                                ((Scoped14)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                    197,
+                                                                    (FactoryDelegate)((IResolverContext r) => //$
+                                                                        new Scoped14()),
+                                                                    r)))),
+                                                        r)),
+                                                    ((ScopedFac13)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                        193,
+                                                        (FactoryDelegate)((IResolverContext r) => //$
+                                                            default(Func<Single1, Scoped14, ScopedFac14, ScopedFac13>)/* (!) Please provide the non-default value for the constant */.Invoke(
+                                                                default(Single1)/* (!) Please provide the non-default value for the constant */,
+                                                                ((Scoped14)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                    197,
+                                                                    (FactoryDelegate)((IResolverContext r) => //$
+                                                                        new Scoped14()),
+                                                                    r)),
+                                                                ((ScopedFac14)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                    203,
+                                                                    default(FactoryDelegate)/* (!) Please provide the non-default value for the constant */,
+                                                                    r)))),
+                                                        r)),
+                                                    new Trans13(
+                                                        default(Single14)/* (!) Please provide the non-default value for the constant */,
+                                                        new Trans14()),
+                                                    default(Single1)/* (!) Please provide the non-default value for the constant */,
+                                                    default(SingleObj1)/* (!) Please provide the non-default value for the constant */)),
+                                            r)))),
+                                r)),
+                            ((Scoped2)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                166,
+                                (FactoryDelegate)((IResolverContext r) => //$
+                                    new Scoped2(
+                                        default(Single22)/* (!) Please provide the non-default value for the constant */,
+                                        default(SingleObj22)/* (!) Please provide the non-default value for the constant */,
+                                        ((ScopedFac22)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                            184,
+                                            (FactoryDelegate)((IResolverContext r) => //$
+                                                default(Func<Scoped23, Single2, SingleObj23, ScopedFac22>)/* (!) Please provide the non-default value for the constant */.Invoke(
+                                                    ((Scoped23)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                        188,
+                                                        (FactoryDelegate)((IResolverContext r) => //$
+                                                            new Scoped23(
+                                                                default(Single2)/* (!) Please provide the non-default value for the constant */,
+                                                                ((Scoped24)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                    198,
+                                                                    (FactoryDelegate)((IResolverContext r) => //$
+                                                                        new Scoped24()),
+                                                                    r)))),
+                                                        r)),
+                                                    default(Single2)/* (!) Please provide the non-default value for the constant */,
+                                                    default(SingleObj23)/* (!) Please provide the non-default value for the constant */)),
+                                            r)),
+                                        new Trans22(
+                                            new Trans23(
+                                                default(Single24)/* (!) Please provide the non-default value for the constant */,
+                                                new Trans24()),
+                                            default(Single23)/* (!) Please provide the non-default value for the constant */,
+                                            default(SingleObj23)/* (!) Please provide the non-default value for the constant */),
+                                        default(Single2)/* (!) Please provide the non-default value for the constant */,
+                                        default(SingleObj2)/* (!) Please provide the non-default value for the constant */,
+                                        ((Scoped22)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                            178,
+                                            (FactoryDelegate)((IResolverContext r) => //$
+                                                new Scoped22(
+                                                    default(Single23)/* (!) Please provide the non-default value for the constant */,
+                                                    default(SingleObj23)/* (!) Please provide the non-default value for the constant */,
+                                                    ((Scoped23)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                        188,
+                                                        (FactoryDelegate)((IResolverContext r) => //$
+                                                            new Scoped23(
+                                                                default(Single2)/* (!) Please provide the non-default value for the constant */,
+                                                                ((Scoped24)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                    198,
+                                                                    (FactoryDelegate)((IResolverContext r) => //$
+                                                                        new Scoped24()),
+                                                                    r)))),
+                                                        r)),
+                                                    ((ScopedFac23)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                        194,
+                                                        (FactoryDelegate)((IResolverContext r) => //$
+                                                            default(Func<Single2, Scoped24, ScopedFac24, ScopedFac23>)/* (!) Please provide the non-default value for the constant */.Invoke(
+                                                                default(Single2)/* (!) Please provide the non-default value for the constant */,
+                                                                ((Scoped24)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                    198,
+                                                                    (FactoryDelegate)((IResolverContext r) => //$
+                                                                        new Scoped24()),
+                                                                    r)),
+                                                                ((ScopedFac24)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                    204,
+                                                                    default(FactoryDelegate)/* (!) Please provide the non-default value for the constant */,
+                                                                    r)))),
+                                                        r)),
+                                                    new Trans23(
+                                                        default(Single24)/* (!) Please provide the non-default value for the constant */,
+                                                        new Trans24()),
+                                                    default(Single2)/* (!) Please provide the non-default value for the constant */,
+                                                    default(SingleObj2)/* (!) Please provide the non-default value for the constant */)),
+                                            r)))),
+                                r)),
+                            new Trans1(
+                                new Trans13(
+                                    default(Single14)/* (!) Please provide the non-default value for the constant */,
+                                    new Trans14()),
+                                new Trans23(
+                                    default(Single24)/* (!) Please provide the non-default value for the constant */,
+                                    new Trans24()),
+                                default(Single13)/* (!) Please provide the non-default value for the constant */,
+                                default(Single1)/* (!) Please provide the non-default value for the constant */,
+                                default(SingleObj1)/* (!) Please provide the non-default value for the constant */),
+                            new Trans2(
+                                new Trans13(
+                                    default(Single14)/* (!) Please provide the non-default value for the constant */,
+                                    new Trans14()),
+                                new Trans23(
+                                    default(Single24)/* (!) Please provide the non-default value for the constant */,
+                                    new Trans24()),
+                                default(Single23)/* (!) Please provide the non-default value for the constant */,
+                                default(Single2)/* (!) Please provide the non-default value for the constant */,
+                                default(SingleObj2)/* (!) Please provide the non-default value for the constant */),
+                            ((ScopedFac1)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                171,
+                                (FactoryDelegate)((IResolverContext r) => //$
+                                    default(Func<Scoped1, Scoped3, Single1, SingleObj1, ScopedFac1>)/* (!) Please provide the non-default value for the constant */.Invoke(
+                                        ((Scoped1)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                            165,
+                                            (FactoryDelegate)((IResolverContext r) => //$
+                                                new Scoped1(
+                                                    default(Single12)/* (!) Please provide the non-default value for the constant */,
+                                                    default(SingleObj12)/* (!) Please provide the non-default value for the constant */,
+                                                    ((ScopedFac12)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                        183,
+                                                        (FactoryDelegate)((IResolverContext r) => //$
+                                                            default(Func<Scoped13, Single1, SingleObj13, ScopedFac12>)/* (!) Please provide the non-default value for the constant */.Invoke(
+                                                                ((Scoped13)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                    187,
+                                                                    (FactoryDelegate)((IResolverContext r) => //$
+                                                                        new Scoped13(
+                                                                            default(Single1)/* (!) Please provide the non-default value for the constant */,
+                                                                            ((Scoped14)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                                197,
+                                                                                (FactoryDelegate)((IResolverContext r) => //$
+                                                                                    new Scoped14()),
+                                                                                r)))),
+                                                                    r)),
+                                                                default(Single1)/* (!) Please provide the non-default value for the constant */,
+                                                                default(SingleObj13)/* (!) Please provide the non-default value for the constant */)),
+                                                        r)),
+                                                    new Trans12(
+                                                        new Trans13(
+                                                            default(Single14)/* (!) Please provide the non-default value for the constant */,
+                                                            new Trans14()),
+                                                        default(Single13)/* (!) Please provide the non-default value for the constant */,
+                                                        default(SingleObj13)/* (!) Please provide the non-default value for the constant */),
+                                                    default(Single1)/* (!) Please provide the non-default value for the constant */,
+                                                    default(SingleObj1)/* (!) Please provide the non-default value for the constant */,
+                                                    ((Scoped12)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                        177,
+                                                        (FactoryDelegate)((IResolverContext r) => //$
+                                                            new Scoped12(
+                                                                default(Single13)/* (!) Please provide the non-default value for the constant */,
+                                                                default(SingleObj13)/* (!) Please provide the non-default value for the constant */,
+                                                                ((Scoped13)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                    187,
+                                                                    (FactoryDelegate)((IResolverContext r) => //$
+                                                                        new Scoped13(
+                                                                            default(Single1)/* (!) Please provide the non-default value for the constant */,
+                                                                            ((Scoped14)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                                197,
+                                                                                (FactoryDelegate)((IResolverContext r) => //$
+                                                                                    new Scoped14()),
+                                                                                r)))),
+                                                                    r)),
+                                                                ((ScopedFac13)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                    193,
+                                                                    (FactoryDelegate)((IResolverContext r) => //$
+                                                                        default(Func<Single1, Scoped14, ScopedFac14, ScopedFac13>)/* (!) Please provide the non-default value for the constant */.Invoke(
+                                                                            default(Single1)/* (!) Please provide the non-default value for the constant */,
+                                                                            ((Scoped14)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                                197,
+                                                                                (FactoryDelegate)((IResolverContext r) => //$
+                                                                                    new Scoped14()),
+                                                                                r)),
+                                                                            ((ScopedFac14)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                                203,
+                                                                                default(FactoryDelegate)/* (!) Please provide the non-default value for the constant */,
+                                                                                r)))),
+                                                                    r)),
+                                                                new Trans13(
+                                                                    default(Single14)/* (!) Please provide the non-default value for the constant */,
+                                                                    new Trans14()),
+                                                                default(Single1)/* (!) Please provide the non-default value for the constant */,
+                                                                default(SingleObj1)/* (!) Please provide the non-default value for the constant */)),
+                                                        r)))),
+                                            r)),
+                                        ((Scoped3)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                            175,
+                                            (FactoryDelegate)((IResolverContext r) => //$
+                                                new Scoped3()),
+                                            r)),
+                                        default(Single1)/* (!) Please provide the non-default value for the constant */,
+                                        default(SingleObj1)/* (!) Please provide the non-default value for the constant */)),
+                                r)),
+                            ((ScopedFac2)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                172,
+                                (FactoryDelegate)((IResolverContext r) => //$
+                                    default(Func<Scoped2, Scoped4, Single2, SingleObj2, ScopedFac2>)/* (!) Please provide the non-default value for the constant */.Invoke(
+                                        ((Scoped2)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                            166,
+                                            (FactoryDelegate)((IResolverContext r) => //$
+                                                new Scoped2(
+                                                    default(Single22)/* (!) Please provide the non-default value for the constant */,
+                                                    default(SingleObj22)/* (!) Please provide the non-default value for the constant */,
+                                                    ((ScopedFac22)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                        184,
+                                                        (FactoryDelegate)((IResolverContext r) => //$
+                                                            default(Func<Scoped23, Single2, SingleObj23, ScopedFac22>)/* (!) Please provide the non-default value for the constant */.Invoke(
+                                                                ((Scoped23)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                    188,
+                                                                    (FactoryDelegate)((IResolverContext r) => //$
+                                                                        new Scoped23(
+                                                                            default(Single2)/* (!) Please provide the non-default value for the constant */,
+                                                                            ((Scoped24)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                                198,
+                                                                                (FactoryDelegate)((IResolverContext r) => //$
+                                                                                    new Scoped24()),
+                                                                                r)))),
+                                                                    r)),
+                                                                default(Single2)/* (!) Please provide the non-default value for the constant */,
+                                                                default(SingleObj23)/* (!) Please provide the non-default value for the constant */)),
+                                                        r)),
+                                                    new Trans22(
+                                                        new Trans23(
+                                                            default(Single24)/* (!) Please provide the non-default value for the constant */,
+                                                            new Trans24()),
+                                                        default(Single23)/* (!) Please provide the non-default value for the constant */,
+                                                        default(SingleObj23)/* (!) Please provide the non-default value for the constant */),
+                                                    default(Single2)/* (!) Please provide the non-default value for the constant */,
+                                                    default(SingleObj2)/* (!) Please provide the non-default value for the constant */,
+                                                    ((Scoped22)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                        178,
+                                                        (FactoryDelegate)((IResolverContext r) => //$
+                                                            new Scoped22(
+                                                                default(Single23)/* (!) Please provide the non-default value for the constant */,
+                                                                default(SingleObj23)/* (!) Please provide the non-default value for the constant */,
+                                                                ((Scoped23)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                    188,
+                                                                    (FactoryDelegate)((IResolverContext r) => //$
+                                                                        new Scoped23(
+                                                                            default(Single2)/* (!) Please provide the non-default value for the constant */,
+                                                                            ((Scoped24)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                                198,
+                                                                                (FactoryDelegate)((IResolverContext r) => //$
+                                                                                    new Scoped24()),
+                                                                                r)))),
+                                                                    r)),
+                                                                ((ScopedFac23)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                    194,
+                                                                    (FactoryDelegate)((IResolverContext r) => //$
+                                                                        default(Func<Single2, Scoped24, ScopedFac24, ScopedFac23>)/* (!) Please provide the non-default value for the constant */.Invoke(
+                                                                            default(Single2)/* (!) Please provide the non-default value for the constant */,
+                                                                            ((Scoped24)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                                198,
+                                                                                (FactoryDelegate)((IResolverContext r) => //$
+                                                                                    new Scoped24()),
+                                                                                r)),
+                                                                            ((ScopedFac24)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                                                                204,
+                                                                                default(FactoryDelegate)/* (!) Please provide the non-default value for the constant */,
+                                                                                r)))),
+                                                                    r)),
+                                                                new Trans23(
+                                                                    default(Single24)/* (!) Please provide the non-default value for the constant */,
+                                                                    new Trans24()),
+                                                                default(Single2)/* (!) Please provide the non-default value for the constant */,
+                                                                default(SingleObj2)/* (!) Please provide the non-default value for the constant */)),
+                                                        r)))),
+                                            r)),
+                                        ((Scoped4)ResolverContext.GetCurrentScopeOrThrow(r).GetOrAddViaFactoryDelegate(
+                                            176,
+                                            (FactoryDelegate)((IResolverContext r) => //$
+                                                new Scoped4()),
+                                            r)),
+                                        default(Single2)/* (!) Please provide the non-default value for the constant */,
+                                        default(SingleObj2)/* (!) Please provide the non-default value for the constant */)),
+                                r)),
+                            default(SingleObj1)/* (!) Please provide the non-default value for the constant */,
+                            default(SingleObj2)/* (!) Please provide the non-default value for the constant */)),
+                    r));
+        }
+    }
+}
